Extract `test_casing` types into module
